functional mouseover charts
