+++ packages/elements/stories/Line.stories.js
@@ -2,7 +2,7 @@
 import { withKnobs, select, number, text } from "@storybook/addon-knobs";
 import BasicLine from "./Line/Basic.svelte";
 import Line01Story from "./Line/Line01.svelte";
-import Line03Story from "./Line/Line03.svelte";
+import MetricsDashboardStory from "./Line/metrics-dashboard/Dashboard.svelte";
 import Line04Story from "./Line/Line04.svelte";
 import Line05Story from "./Line/Line05.svelte";
 
@@ -32,8 +32,8 @@ export const Line01 = () => ({
   Component: Line01Story,
 });
 
-export const Line03 = () => ({
-  Component: Line03Story,
+export const MetricsDashboard = () => ({
+  Component: MetricsDashboardStory,
 });
 
 export const Line04 = () => ({
